+++ CHANGELOG.md
@@ -74,8 +74,7 @@ Main (unreleased)
   now export receivers immediately, removing the need for dependant components
   to be evaluated twice at process startup. (@rfratto)
 
-- Add missing setting to configure instance key, extra labels and autoscrape for
-  Eventhandler integration. (@marctc)
+- Add missing setting to configure instance key for Eventhandler integration. (@marctc)
 
 ### Bugfixes